Added Command file to clone any repo
This is a super useful tool I use to clone any repo hosted on GitHub.

It will first ask for the GitHub repo's username. Then it will ask for the specific repo name you are looking for.

It will create a folder in the directory where this command file is run from, and clone into it.